reversing the trinity test
