--- frontend/src/state/selectors.js
@@ -1,6 +1,0 @@
-export const loginErrorMessageSelector = (state) => state.user.loginError;
-export const signupErrorMessageSelector = (state) => state.user.signupError;
-export const usernameSelector = (state) => state.user.username;
-export const accessTokenSelector = (state) => state.user.accessToken;
-export const teamsSelector = (state) => state.user.teams;
-export const emailSelector = (state) => state.user.email;
